Break out base client, add reports client
BaseClient contains the shared functionality for all MWS service calls,
and PaymentsClient has all the public methods from the former Client
class.  New class ReportsClient allows calls to getReportList and
getReport.

TODO: support more parameters for getReportList
